Hack to prevent pry-byebug from hanging on find during a binding.pry
Hack to remove a line that causes to pry.byebug to hang during a binding.pry when find is used. This line is needed for jruby, but shouldn't be needed for mri. So this is a temporary work around until a fix is found in byebug which is the true culprit.